fix: 부적절한 위치의 Dto 원상복구
