+++ src/test/java/chess/controller/WebChessControllerTest.java
@@ -23,10 +23,10 @@ import chess.dto.BoardsDto;
 import chess.dto.request.MoveRequestDto;
 import chess.dto.request.RoomRequestDto;
 import chess.dto.response.ErrorResponseDto;
-import chess.dto.response.ErrorResponseDto.StatusResponseDto;
 import chess.dto.response.GameResponseDto;
 import chess.dto.response.RoomResponseDto;
 import chess.dto.response.RoomsResponseDto;
+import chess.dto.response.StatusResponseDto;
 import chess.entity.BoardEntity;
 import chess.entity.RoomEntity;
 import chess.exception.RoomNotFoundException;
